Extract restaurant validation to an util class
